Made ExpiresAfterSellIn an optional parameter that defaults to false; added a test to prove it out, and changed some test/test method names to reflect changes to the underlying objects
