+++ GildedRoseTests/ItemProcessingRuleTests.cs
@@ -1,46 +1,59 @@
 ﻿using GildedRoseKata;
-using System.Collections.Generic;
-using System.Data;
 using Xunit;
 
 namespace GildedRoseTests
 {
-    public class ItemProcessingRulesTests
+    public class ItemProcessingRuleTests
     {
         [Fact]
-        public void GetRulesForItem_WhenGivenItem_ReturnsCorrectRuleForItem()
+        public void NoUpdateRule_RequiresQualityUpdate_SetToFalse()
         {
-            ItemProcessingRule agedBrie = new(1, false);
-            NoUpdateRule sulfuras = new();
-            Dictionary<string, IItemProcessingRules> rules = new()
-            {
-                { "Sulfuras", sulfuras },
-                { "Aged Brie", agedBrie }
-            };
-
-            ItemProcessingRuleProvider processingRules = new(rules, new ItemProcessingRule(-1, false));
-            var rule = processingRules.GetRuleForItem(new Item { Name = "Sulfuras"});
-
-            Assert.Equal(sulfuras, rule);
+            var rule = new NoUpdateRule();
+            Assert.False(rule.RequiresQualityUpdate);
         }
 
         [Fact]
-        public void GetRulesForItem_WhenGivenItemAndNoMatchingRuleForItem_ReturnsDefaultRule()
+        public void ItemProcessingRule_RequiresQualityUpdate_SetToTrue()
         {
-            ItemProcessingRule agedBrie = new (1, false);
-            NoUpdateRule sulfuras = new ();
-            ItemProcessingRule defaultRule = new(-1, false);
+            var rule = new ItemProcessingRule(-1);
+            Assert.True(rule.RequiresQualityUpdate);
+        }
 
-            Dictionary<string, IItemProcessingRules> rules = new()
-            {
-                { "Sulfuras", sulfuras },
-                { "Aged Brie", agedBrie }
-            };
+        [Fact]
+        public void ItemProcessingRule_ExpiresAfterSellIn_SetToFalse()
+        {
+            var rule = new ItemProcessingRule(-1);
+            Assert.False(rule.ExpiresAfterSellIn);
+        }
 
-            ItemProcessingRuleProvider processingRules = new(rules, defaultRule);
-            var rule = processingRules.GetRuleForItem(new Item { Name = "Backstage passes" });
+        [Fact]
+        public void GetExtraQualityAdjustmentBySellIn_SingleSellInThreshold_ItemInsideThreshold_ReturnsCorrectQualityAdjustment()
+        {
+            var rule = new ItemProcessingRule(-1, sellInQualityAdjustmentRules: new SellInQualityAdjustmentRule() { SellInThreshold = 10, QualityAdjustment = 2 });
+            int extraQualityAdjustment = rule.GetExtraQualityAdjustmentBySellIn(5);
 
-            Assert.Equal(defaultRule, rule);
+            Assert.Equal(2, extraQualityAdjustment);
+        }
+
+
+        [Fact]
+        public void GetExtraQualityAdjustmentBySellIn_SingleSellInThreshold_ItemOutsideThreshold_ReturnsCorrectQualityAdjustment()
+        {
+            var rule = new ItemProcessingRule(-1, sellInQualityAdjustmentRules: new SellInQualityAdjustmentRule() { SellInThreshold = 10, QualityAdjustment = 2 });
+            int extraQualityAdjustment = rule.GetExtraQualityAdjustmentBySellIn(11);
+
+            Assert.Equal(0, extraQualityAdjustment);
+        }
+
+        [Fact]
+        public void GetExtraQualityAdjustmentBySellIn_DoubleSellInThreshold_ItemInsideInnerThreshold_ReturnsSumQualityAdjustment()
+        {
+            var rule = new ItemProcessingRule(-1, sellInQualityAdjustmentRules: [
+                new SellInQualityAdjustmentRule() { SellInThreshold = 10, QualityAdjustment = 1 },
+                new SellInQualityAdjustmentRule() { SellInThreshold = 5, QualityAdjustment = 1 }]);
+            int extraQualityAdjustment = rule.GetExtraQualityAdjustmentBySellIn(4);
+
+            Assert.Equal(2, extraQualityAdjustment);
         }
     }
-}  
+}
